k, so please read very carefully as it may be confusing(but the changes are necessary)
the old Solver.java, is now the Board.java
and old PlaySolver.java is now the Solver.java

As the code Solver setup and create the board as a board object, it will be very confusing for the reader and our document if we kept the old naming system.

I also added a block class which allow us to check if the current configure is the same as the final configuration w/o doing crazy check on the block_map

So now, the board.java can handle adding_blocks, display_board, check_finalconfig. (2 out of 3 required tasks are done. 

Hey Emin: Thanks for the input, but I'm not sure we can use your scheme, coz requirement stated: "implementation of a good hashing scheme", and O(n^2) may not be a very good scheme for hashing.

email me if there are questions, I need to study for my midterm, so I don't know if I can finish the 3rd task, I will try.............